Move UpdatedMeta to a target
